+++ packages/goat/src/methods/debug/timeFunction.js
@@ -1,0 +1,7 @@
+function timeFunction(fn, string) {
+  console.time(string);
+  fn();
+  console.timeEnd(string);
+}
+
+module.exports = timeFunction;