+++ packages/goat/src/validators/schema.js
@@ -1,4 +1,3 @@
-const { validate } = require('jsonschema');
 const Notifier = require('../methods/notifications/notifyHandler');
 
 /**
@@ -7,7 +6,8 @@ const Notifier = require('../methods/notifications/notifyHandler');
  * @param {Object} schema
  * @returns {Boolean} isValid
  */
-const checkSchema = (config, schema) => {
+function checkSchema (config, schema) {
+  const { validate } = require('jsonschema');
   const result = validate(config, schema);
   if (result.errors.length === 0) {
     return true;