+++ packages/goat/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.5.1
+- Feature: added `project ls` command
+- Feature: added a debug flag
+- Fix: show descriptions of build-in modules
+
 ## 1.5.0
 - Feature: Added module system, 
 - Performance: reduced load times by optimizing module loading